fix(core): :zap: Fix elements reversing rendering issue and Remove the rendering from 'html' property
Remove the old rendering from 'html' , the old way to render was save the (strings,number..) children to the html then save the  component children to the children array and render the html then the children , and this introduced that you'll JSX won't be like you output  and your texts .. will be rendered and then the children. we removed the html and make all expression and children in one array. we removed some types and changed many files.